requirements.txt
langchain
openai
pinecone-client==2.2.4
streamlit
pypdf
tiktoken
cohere
